+++ src/styles/RoomList.module.css
@@ -1,0 +1,10 @@
+.card {
+  width: 240px;
+}
+
+.container {
+  margin: auto;
+  gap: 64px;
+  position: relative;
+  overflow-x: scroll;
+}
