+++ src/styles/elements.css
@@ -51,10 +51,6 @@ tr:nth-child(odd) {
   background-color: var(--gray);
 }
 
-.card {
-  width: 240px;
-}
-
 @media screen and (min-width: 992px) {
   .info-table {
     width: 300px;
