+++ src/styles/layout.css
@@ -51,6 +51,7 @@
   height: 100%;
   width: 100%;
   overflow-y: auto;
+  overflow-x: hidden;
 }
 
 .open-nav {
